dtnorm.c passes unit test in test/xdtnorm.c.
The unit test examines several sets of parameter values so as to exercise all 4
of the underlying algorithms. For each set of parameter values, it calculates
two statistics: the mean, and the fraction of observations in the left half of
the interval. To pass the test, these statistics must be near their expectated values.